Move 'operation' into 'operations' subfolder.
