+++ earthmover/operations/dataframe.py
@@ -3,7 +3,7 @@ import numpy as np
 import pandas as pd
 
 from earthmover.node import Node
-from earthmover.nodes.operation import Operation
+from earthmover.operations.operation import Operation
 
 from typing import Dict, List, Tuple
 from typing import TYPE_CHECKING
